[UPDATE] merapihkan structure code, menambahkan people dan syllabus
